Implement simple copy/remove
This was necessary to perform moves across devices under linux.
The current implementation is still rough around the edges and needs
better error handling. For now a cross-device move prints the
corresponding error message. This will be fix as soon as I find out
which other errors could be caused by os.Rename.